Move 'delete component' CLI files to a dedicated 'cli/delete/component' package
