+++ pkg/odo/cli/delete/component/component_test.go
@@ -1,4 +1,4 @@
-package delete
+package component
 
 import (
 	"errors"
